backend/requirements.txt
# Django Core
Django==5.2.9
djangorestframework==3.15.2
django-filter==24.3

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter

# Caching & Task Queue
redis==5.2.1
django-redis==5.4.0
celery==5.4.0

# Image Processing
Pillow==11.0.0

# CORS
django-cors-headers==4.6.0

# Authentication
djangorestframework-simplejwt==5.4.0

# Environment Variables
python-decouple==3.8

# Utilities
python-slugify==8.0.4
